refactor directories and concepts
